デバッグ修正。Freezeを定義する前にreloadしていたところを修正。会話進行について、エンティティクリックと、アドベンチャーモード時のブロッククリックに対応、クリックイベントをリファクタリング、その他修正

+++ src/main/java/seisyun/rpgtext/CustomConfig.java
@@ -48,6 +48,10 @@ public class CustomConfig {
         this.file = fileName;
         makeDirectory(directory);
         configFile = new File(directory, this.file);
+        // ファイルが既に存在している場合は作成しない
+        if(configFile.exists()){
+            return;
+        }
         try{
             Files.copy(plugin.getResource(resource),configFile.toPath());
             plugin.getLogger().info("Created a \"" +  fileName + "\" file.");
